+++ CHANGELOG.md
@@ -23,6 +23,8 @@ Types of changes:
 - Add goshy as bash script and NixOS program, [@riotbib](https://github.com/riotbib) in [#27](https://github.com/oxzi/gosh/pull/27).
 - Created Store RPC working on Unix domain sockets to allow a `fork`+`exec`ed daemon.
 - Configuration through YAML configuration file.
+- Configurable index template and static files, partially by [@riotbib](https://github.com/riotbib) in [#45](https://github.com/oxzi/gosh/pull/45).
+- ID of new items is now configurable both in length as well as in source (random, wordlist).
 
 ### Changed
 - Dependency version bumps.
@@ -30,7 +32,8 @@ Types of changes:
 - `goshd` became `gosh`.
 - Made `gosh` a `chroot`ed, privilege dropped, `fork`+`exec`ed daemon.
 - OpenBSD installation changed due to structural program changes.
-- Extract web template into a more editable file, [@riotbib](https://github.com/riotbib) in [#45](https://github.com/oxzi/gosh/pull/45).
+- Bumped required Go version from 1.19 to 1.21.
+- Replaced logrus logging with Go's new `log/slog` and do wrapping for child processes.
 
 ### Deprecated
 ### Removed